Implement leave a review feature

+++ shop/shop_data/src/main/java/com/eshop/shop_data/remote/request/ReviewRequest.kt
@@ -1,3 +1,3 @@
 package com.eshop.shop_data.remote.request
 
-data class ReviewRequest(val comment: String, val rating: Double)
+data class ReviewRequest(val comment: String, val rating: Int)
